Rename capture details file
