+++ src/components/capture/EditorAndLogs.tsx
