+++ src/components/tables/DetailsPanel.tsx
@@ -1,5 +1,5 @@
 import { Collapse, TableCell, TableRow } from '@mui/material';
-import CaptureDetails from 'components/capture/Details';
+import CaptureDetails from 'components/capture/EditorAndLogs';
 import { createEditorStore } from 'components/editor/Store';
 import { ZustandProvider } from 'hooks/useZustand';
 import { ENTITY } from 'types';
